v1/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.7
python-multipart>=0.0.6

# Document processing
pytesseract>=0.3.10
Pillow>=9.5.0
pdf2image>=1.16.3
PyPDF2>=3.0.1
openai>=0.27.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Original dependency
agentic-doc==0.3.1